requirements.txt
grpcio==1.68.1
grpcio-tools==1.68.1
owlready2==0.47
protobuf==5.29.2
setuptools==75.6.0
